+++ NEWS.md
@@ -1,3 +1,9 @@
+# phenoptrReports 0.3.1.9000
+
+- Update tests to work with testthat 3.1.
+- Fix spatial map viewer to work with older composite images 
+  that are missing ImageDescription (#52)
+
 # phenoptrReports 0.3.1
 **2021-08-24**